[FreeMotion] Add option to update scene components after BeginEvent
This is an experimental feature to allow an automatic update of scene
components in a scene using the FreeMotionAnimationLoop, right after the
end of the animateBeginEvent, and before the solution of the mechanical
system. This update is simply done by calling
MechanicalOperations::propagateXandV, which in turns launches a
MechanicalPropagateOnlyPositionAndVelocityVisitor. This visitor will,
namely, go through the scene mappings and call the apply() method.

In order to introduce this change while not breaking existing scenes
behaviour, we add a bool data in the FreeMotionAnimatonLoop, setting the
data default value to False. The call to propagateXandV will be made
after animateBeginEvent only if a scene user specifically sets this data
to True.

The typical use of this feature is to update the scene if Mappings were
dynamically modified (for instance by Python scripting) during  the
animateBeginEvent phase, and that the output MechanicalObjects of those
mappings have to be updated before entering the system solution
pipeline.

A different, less invasive approach regarding the
FreeMotionAnimationLoop, would be to implement a binding in order for a
user to call the propagateXandV method (or launch the corresponding
visitor) directly from a Python script.